+++ pkg/cmd/delete_credential.go
@@ -1,8 +1,10 @@
 package cmd
 
 import (
+	"errors"
 	"fmt"
 	"io"
+	"reflect"
 
 	"github.com/spf13/cobra"
 
@@ -11,6 +13,10 @@ import (
 	"github.com/ZupIT/ritchie-cli/pkg/prompt"
 	"github.com/ZupIT/ritchie-cli/pkg/stdin"
 )
+
+type inputDeleteCredential struct {
+	provider string
+}
 
 // deleteCredentialCmd type for set credential command
 type deleteCredentialCmd struct {
@@ -24,6 +30,15 @@ type deleteCredentialCmd struct {
 // deleteCredential type for stdin json decoder
 type deleteCredential struct {
 	Provider string `json:"provider"`
+}
+
+var deleteCredentialFlags = flags{
+	{
+		name:        providerFlagName,
+		kind:        reflect.String,
+		defValue:    "",
+		description: "Provider name to delete",
+	},
 }
 
 // NewDeleteCredentialCmd creates a new cmd instance
@@ -46,49 +61,32 @@ func NewDeleteCredentialCmd(
 		Use:       "credential",
 		Short:     "Delete credential",
 		Long:      `Delete credential from current env`,
-		RunE:      RunFuncE(s.runStdin(), s.runPrompt()),
+		RunE:      RunFuncE(s.runStdin(), s.runFormula()),
 		ValidArgs: []string{""},
 		Args:      cobra.OnlyValidArgs,
 	}
-	cmd.LocalFlags()
+
+	addReservedFlags(cmd.Flags(), deleteCredentialFlags)
+
 	return cmd
 }
 
-func (d deleteCredentialCmd) runPrompt() CommandRunnerFunc {
+func (d deleteCredentialCmd) runFormula() CommandRunnerFunc {
 	return func(cmd *cobra.Command, args []string) error {
-		env, err := d.currentEnv()
+		curEnv, err := d.currentEnv()
 		if err != nil {
 			return err
 		}
-		prompt.Info(fmt.Sprintf("Current env: %s", env))
+		prompt.Info(fmt.Sprintf("Current env: %s", curEnv))
 
-		data, err := d.ReadCredentialsValueInEnv(d.CredentialsPath(), env)
+		inputParams, err := d.resolveInput(cmd, curEnv)
 		if err != nil {
 			return err
-		}
-
-		if len(data) <= 0 {
-			prompt.Error("You have no defined credentials in this env")
+		} else if inputParams.provider == "" {
 			return nil
 		}
 
-		var providers []string
-		for _, c := range data {
-			providers = append(providers, c.Provider)
-		}
-
-		cred, err := d.List("Credentials: ", providers)
-		if err != nil {
-			return err
-		}
-
-		if b, err := d.Bool("Are you sure want to delete this credential?", []string{"yes", "no"}); err != nil {
-			return err
-		} else if !b {
-			return nil
-		}
-
-		if err := d.Delete(cred); err != nil {
+		if err := d.Delete(inputParams.provider); err != nil {
 			return err
 		}
 
@@ -97,6 +95,53 @@ func (d deleteCredentialCmd) runPrompt() CommandRunnerFunc {
 	}
 }
 
+func (d *deleteCredentialCmd) resolveInput(cmd *cobra.Command, context string) (inputDeleteCredential, error) {
+	if IsFlagInput(cmd) {
+		return d.resolveFlags(cmd)
+	}
+	return d.resolvePrompt(context)
+}
+
+func (d *deleteCredentialCmd) resolvePrompt(context string) (inputDeleteCredential, error) {
+	data, err := d.ReadCredentialsValueInEnv(d.CredentialsPath(), context)
+	if err != nil {
+		return inputDeleteCredential{}, err
+	}
+
+	if len(data) == 0 {
+		return inputDeleteCredential{}, errors.New("you have no defined credentials in this env")
+	}
+
+	providers := make([]string, len(data))
+	for _, c := range data {
+		providers = append(providers, c.Provider)
+	}
+
+	provider, err := d.List("Credentials: ", providers)
+	if err != nil {
+		return inputDeleteCredential{}, err
+	}
+
+	if b, err := d.Bool("Are you sure want to delete this credential?", []string{"yes", "no"}); err != nil {
+		return inputDeleteCredential{}, err
+	} else if !b {
+		return inputDeleteCredential{}, nil
+	}
+	return inputDeleteCredential{provider}, nil
+}
+
+func (d *deleteCredentialCmd) resolveFlags(cmd *cobra.Command) (inputDeleteCredential, error) {
+	provider, err := cmd.Flags().GetString(providerFlagName)
+	if err != nil {
+		return inputDeleteCredential{}, err
+	} else if provider == "" {
+		return inputDeleteCredential{}, errors.New("please provide a value for 'provider'")
+	}
+
+	return inputDeleteCredential{provider}, nil
+}
+
+// TODO: remove upon stdin deprecation
 func (d deleteCredentialCmd) runStdin() CommandRunnerFunc {
 	return func(cmd *cobra.Command, args []string) error {
 		dc, err := d.stdinResolver(cmd.InOrStdin())
@@ -104,12 +149,12 @@ func (d deleteCredentialCmd) runStdin() CommandRunnerFunc {
 			return err
 		}
 
-		env, err := d.currentEnv()
+		curEnv, err := d.currentEnv()
 		if err != nil {
 			return err
 		}
 
-		data, err := d.ReadCredentialsValueInEnv(d.CredentialsPath(), env)
+		data, err := d.ReadCredentialsValueInEnv(d.CredentialsPath(), curEnv)
 		if err != nil {
 			return err
 		}